Split modules to keep proxy_service.rs concise
