+++ src/proxy_service.rs
@@ -18,25 +18,12 @@
 use std::sync::Arc;
 
 use anyhow::{Context, anyhow};
-use arrow::datatypes::{Schema, SchemaRef};
 use arrow_flight::{
     FlightDescriptor,
-    FlightEndpoint,
     FlightInfo,
     Ticket,
-    encode::FlightDataEncoderBuilder,
-    error::FlightError,
     flight_service_server::FlightServiceServer,
-    sql::{ProstMessageExt, TicketStatementQuery},
 };
-
-use datafusion::{
-    logical_expr::LogicalPlan,
-    physical_plan::{ExecutionPlan, Partitioning, coalesce_partitions::CoalescePartitionsExec, displayable},
-    prelude::SessionContext,
-};
-
-use futures::TryStreamExt;
 use parking_lot::Mutex;
 use prost::Message;
 use tokio::{
@@ -46,420 +33,27 @@ use tokio::{
 use tonic::{Request, Response, Status, async_trait, transport::Server};
 
 use crate::{
-    explain::{DistributedExplainExec, is_explain_query},
-    protobuf::DistributedExplainExecNode,
+    explain::is_explain_query,
     flight::{FlightSqlHandler, FlightSqlServ},
+    flight_handlers::FlightRequestHandler,
     k8s::get_worker_addresses,
     logging::{debug, info, trace},
-    planning::{
-        add_ctx_extentions,
-        distribute_stages,
-        execution_planning,
-        get_ctx,
-        logical_planning,
-        physical_planning,
-        DFRayStage,
-    },
     protobuf::TicketStatementData,
+    query_planner::QueryPlanner,
     result::Result,
-    stage_reader::DFRayStageReaderExec,
-    util::{display_plan_with_partition_counts, get_addrs},
-    vocab::Addrs,
 };
 
-/// Result of base query preparation containing all planning artifacts for both query and its EXPLAIN
-pub struct QueryPlanBase {
-    pub query_id: String,
-    pub session_context: SessionContext,
-    pub logical_plan: LogicalPlan,
-    pub physical_plan: Arc<dyn ExecutionPlan>,
-    pub distributed_plan: Arc<dyn ExecutionPlan>,
-    pub distributed_stages: Vec<DFRayStage>,
+pub struct DfRayProxyHandler {
+    pub flight_handler: FlightRequestHandler,
 }
-
-impl std::fmt::Debug for QueryPlanBase {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.debug_struct("QueryPlanBase")
-            .field("query_id", &self.query_id)
-            .field("session_context", &"<SessionContext>")
-            .field("logical_plan", &self.logical_plan)
-            .field("physical_plan", &format!("<{}>", self.physical_plan.name()))
-            .field("distributed_plan", &format!("<{}>", self.distributed_plan.name()))
-            .field("stages", &format!("<{} stages>", self.distributed_stages.len()))
-            .finish()
-    }
-}
-
-/// Result of query preparation for execution of both query and its EXPLAIN
-#[derive(Debug)]
-pub struct QueryPlan {
-    pub query_id: String,
-    pub worker_addresses: Addrs,
-    pub final_stage_id: u64,
-    pub schema: SchemaRef,
-    pub explain_data: Option<DistributedExplainExec>,
-}
-
-struct DfRayProxyHandler {}
 
 impl DfRayProxyHandler {
     pub fn new() -> Self {
-        // call this function but ignore the results to bootstrap the worker
-        // discovery mechanism
-        get_worker_addresses().expect("Could not get worker addresses upoon startup");
-        Self {}
-    }
-
-    /// Common planning steps shared by both query and its EXPLAIN
-    ///
-    /// Prepare a query by parsing the SQL, planning it, and distributing the
-    /// physical plan into stages that can be executed by workers.
-    async fn prepare_query_base(&self, sql: &str, query_type: &str) -> Result<QueryPlanBase> {
-        debug!("prepare_query_base: {} SQL = {}", query_type, sql);
-
-        let query_id = uuid::Uuid::new_v4().to_string();
-        let ctx = get_ctx().map_err(|e| anyhow!("Could not create context: {e}"))?;
-
-        let logical_plan = logical_planning(sql, &ctx).await?;
-        let physical_plan = physical_planning(&logical_plan, &ctx).await?;
-
-        // divide the physical plan into chunks (stages) that we can distribute to workers
-        let (distributed_plan, distributed_stages) = execution_planning(physical_plan.clone(), 8192, Some(2)).await?;
-
-        Ok(QueryPlanBase {
-            query_id,
-            session_context: ctx,
-            logical_plan,
-            physical_plan,
-            distributed_plan,
-            distributed_stages,
-        })
-    }
-
-    /// Prepare a distributed query
-    pub async fn prepare_query(&self, sql: &str) -> Result<QueryPlan> {
-        let base_result = self.prepare_query_base(sql, "REGULAR").await?;
-
-        if base_result.distributed_stages.is_empty() {
-            return Err(anyhow!("No stages generated for query").into());
+        // call this function to bootstrap the worker discovery mechanism
+        get_worker_addresses().expect("Could not get worker addresses upon startup");
+        Self {
+            flight_handler: FlightRequestHandler::new(QueryPlanner::new()),
         }
-
-        let worker_addrs = get_worker_addresses()?;
-
-        // gather some information we need to send back such that
-        // we can send a ticket to the client
-        let final_stage = &base_result.distributed_stages[base_result.distributed_stages.len() - 1];
-        let schema = Arc::clone(&final_stage.plan.schema());
-        let final_stage_id = final_stage.stage_id;
-
-        // distribute the stages to workers, further dividing them up
-        // into chunks of partitions (partition_groups)
-        let final_workers = distribute_stages(&base_result.query_id, base_result.distributed_stages, worker_addrs).await?;
-
-        Ok(QueryPlan {
-            query_id: base_result.query_id,
-            worker_addresses: final_workers,
-            final_stage_id,
-            schema,
-            explain_data: None,
-        })
-    }
-
-    /// Prepare an EXPLAIN query
-    /// This method only handles EXPLAIN queries (plan only). EXPLAIN ANALYZE queries are handled as regular queries because they need to be executed.
-    pub async fn prepare_explain(&self, sql: &str) -> Result<QueryPlan> {
-        // Validate that this is actually an EXPLAIN query (not EXPLAIN ANALYZE)
-        if !is_explain_query(sql) {
-            return Err(anyhow!("prepare_explain called with non-EXPLAIN query or EXPLAIN ANALYZE query: {}", sql).into());
-        }
-        
-        // Extract the underlying query from the EXPLAIN statement
-        let underlying_query = sql.trim()
-            .strip_prefix("EXPLAIN")
-            .or_else(|| sql.trim().to_uppercase().strip_prefix("EXPLAIN").map(|_| &sql.trim()[7..]))
-            .unwrap_or(sql)
-            .trim();
-        
-        let base_result = self.prepare_query_base(underlying_query, "EXPLAIN").await?;
-
-        // generate the plan strings
-        let logical_plan_string = format!("{}", base_result.logical_plan.display_indent());
-        let physical_plan_string = format!("{}", displayable(base_result.physical_plan.as_ref()).indent(true));
-        let distributed_plan_string = format!("{}", displayable(base_result.distributed_plan.as_ref()).indent(true));
-        let distributed_stages_string = DistributedExplainExec::format_distributed_stages(base_result.distributed_stages.as_slice());
-
-        // create the schema for EXPLAIN results
-        use arrow::datatypes::{DataType, Field, Schema};
-        let schema = Arc::new(Schema::new(vec![
-            Field::new("plan_type", DataType::Utf8, false),
-            Field::new("plan", DataType::Utf8, false),
-        ]));
-
-        // Create explain data
-        let explain_data = DistributedExplainExec::new(
-            Arc::clone(&schema),
-            logical_plan_string,
-            physical_plan_string,
-            distributed_plan_string,
-            distributed_stages_string,
-        );
-
-        // Create dummy addresses for EXPLAIN (no real workers needed)
-        let mut dummy_addrs = std::collections::HashMap::new();
-        let mut partition_addrs = std::collections::HashMap::new();
-        partition_addrs.insert(0u64, vec![("explain_local".to_string(), "local".to_string())]);
-        dummy_addrs.insert(0u64, partition_addrs);
-
-        Ok(QueryPlan {
-            query_id: base_result.query_id,
-            worker_addresses: dummy_addrs,
-            final_stage_id: 0,
-            schema,
-            explain_data: Some(explain_data),
-        })
-    }
-
-    /// Create a FlightInfo response with the given ticket data
-    fn create_flight_info_response(
-        &self,
-        query_id: String,
-        final_addrs: Addrs,
-        final_stage_id: u64,
-        schema: SchemaRef,
-        explain_data: Option<crate::protobuf::DistributedExplainExecNode>,
-    ) -> Result<FlightInfo, Status> {
-        let mut flight_info = FlightInfo::new()
-            .try_with_schema(&schema)
-            .map_err(|e| Status::internal(format!("Could not create flight info {e:?}")))?;
-
-        let ticket_data = TicketStatementData {
-            query_id,
-            stage_id: final_stage_id,
-            stage_addrs: Some(final_addrs.into()),
-            schema: Some(schema.try_into().map_err(|e| {
-                Status::internal(format!("Could not convert schema {e:?}"))
-            })?),
-            explain_data,
-        };
-
-        let ticket = Ticket::new(
-            TicketStatementQuery {
-                statement_handle: ticket_data.encode_to_vec().into(),
-            }
-            .as_any()
-            .encode_to_vec(),
-        );
-
-        let endpoint = FlightEndpoint::new().with_ticket(ticket);
-        flight_info = flight_info.with_endpoint(endpoint);
-
-        Ok(flight_info)
-    }
-
-    /// Handle EXPLAIN query requests by preparing plans for visualization.
-    /// 
-    /// EXPLAIN queries return comprehensive plan information including logical, physical, 
-    /// distributed plan, and execution stages for analysis and debugging purposes.
-    async fn handle_explain_request(&self, query: &str) -> Result<Response<FlightInfo>, Status> {
-        let plans = self
-            .prepare_explain(query)
-            .await
-            .map_err(|e| Status::internal(format!("Could not prepare EXPLAIN query {e:?}")))?;
-
-        debug!("get flight info: EXPLAIN query id {}", plans.query_id);
-
-        let explain_data = plans.explain_data.map(|data| {
-            DistributedExplainExecNode {
-                schema: data.schema().as_ref().try_into().ok(),
-                logical_plan: data.logical_plan().to_string(),
-                physical_plan: data.physical_plan().to_string(),
-                distributed_plan: data.distributed_plan().to_string(),
-                distributed_stages: data.distributed_stages().to_string(),
-            }
-        });
-
-        let flight_info = self.create_flight_info_response(
-            plans.query_id,
-            plans.worker_addresses,
-            plans.final_stage_id,
-            plans.schema,
-            explain_data
-        )?;
-
-        trace!("get_flight_info_statement done for EXPLAIN");
-        Ok(Response::new(flight_info))
-    }
-
-    /// Handle query requests by preparing execution plans and stages.
-    /// 
-    /// Query focus on execution readiness, returning only the essential 
-    /// metadata needed to execute the distributed query plan.
-    async fn handle_query_request(&self, query: &str) -> Result<Response<FlightInfo>, Status> {
-        let query_plan = self
-            .prepare_query(query)
-            .await
-            .map_err(|e| Status::internal(format!("Could not prepare query {e:?}")))?;
-
-        debug!("get flight info: query id {}", query_plan.query_id);
-
-        let flight_info = self.create_flight_info_response(
-            query_plan.query_id,
-            query_plan.worker_addresses,
-            query_plan.final_stage_id,
-            query_plan.schema,
-            None  // Regular queries don't have explain data
-        )?;
-
-        trace!("get_flight_info_statement done");
-        Ok(Response::new(flight_info))
-    }
-
-    /// Handle execution of EXPLAIN statement queries.
-    /// 
-    /// This function does not execute the plan but returns all plans we want to display to the user.
-    async fn handle_explain_statement_execution(
-        &self,
-        tsd: TicketStatementData,
-        remote_addr: &str,
-    ) -> Result<Response<crate::flight::DoGetStream>, Status> {
-        let explain_data = tsd.explain_data.as_ref()
-            .ok_or_else(|| Status::internal("No explain_data in TicketStatementData for EXPLAIN query"))?;
-        
-        let schema: Schema = explain_data
-            .schema
-            .as_ref()
-            .ok_or_else(|| Status::internal("No schema in ExplainData"))?
-            .try_into()
-            .map_err(|e| Status::internal(format!("Cannot convert schema {e}")))?;
-
-        let explain_plan = Arc::new(DistributedExplainExec::new(
-            Arc::new(schema),
-            explain_data.logical_plan.clone(),
-            explain_data.physical_plan.clone(),
-            explain_data.distributed_plan.clone(),
-            explain_data.distributed_stages.clone(),
-        )) as Arc<dyn ExecutionPlan>;
-
-        debug!(
-            "EXPLAIN request for query_id {} from {} explain plan:\n{}",
-            tsd.query_id,
-            remote_addr,
-            display_plan_with_partition_counts(&explain_plan)
-        );
-
-        // Create dummy addresses for EXPLAIN execution
-        let mut dummy_addrs = std::collections::HashMap::new();
-        let mut partition_addrs = std::collections::HashMap::new();
-        partition_addrs.insert(0u64, vec![("explain_local".to_string(), "local".to_string())]);
-        dummy_addrs.insert(0u64, partition_addrs);
-
-        self.execute_plan_and_build_stream(explain_plan, tsd.query_id, dummy_addrs).await
-    }
-
-    /// Handle execution of regular statement queries
-    /// 
-    /// This function executes the plan and returns the results to the client.
-    async fn handle_regular_statement_execution(
-        &self,
-        tsd: TicketStatementData,
-        remote_addr: &str,
-    ) -> Result<Response<crate::flight::DoGetStream>, Status> {
-        let schema: Schema = tsd
-            .schema
-            .as_ref()
-            .ok_or_else(|| Status::internal("No schema in TicketStatementData"))?
-            .try_into()
-            .map_err(|e| Status::internal(format!("Cannot convert schema {e}")))?;
-
-        // Create an Addrs from the final stage information in the tsd
-        let stage_addrs = tsd.stage_addrs.ok_or_else(|| {
-            Status::internal("No stages_addrs in TicketStatementData, cannot proceed")
-        })?;
-
-        let addrs: Addrs = get_addrs(&stage_addrs).map_err(|e| {
-            Status::internal(format!("Cannot get addresses from stage_addrs {e:?}"))
-        })?;
-
-        trace!("calculated addrs: {:?}", addrs);
-
-        // Validate that addrs contains exactly one stage
-        self.validate_single_stage_addrs(&addrs, tsd.stage_id)?;
-
-        let stage_partition_addrs = addrs.get(&tsd.stage_id)
-            .ok_or_else(|| Status::internal(format!("No partition addresses found for stage_id {}", tsd.stage_id)))?;
-        
-        let plan = Arc::new(
-            DFRayStageReaderExec::try_new(
-                Partitioning::UnknownPartitioning(stage_partition_addrs.len()),
-                Arc::new(schema),
-                tsd.stage_id,
-            )
-            // TODO: revisit this to allow for consuming a particular partition
-            .map(|stg| CoalescePartitionsExec::new(Arc::new(stg)))
-            .map_err(|e| Status::internal(format!("Unexpected error {e}")))?,
-        ) as Arc<dyn ExecutionPlan>;
-
-        debug!(
-            "request for query_id {} from {} reader plan:\n{}",
-            tsd.query_id,
-            remote_addr,
-            display_plan_with_partition_counts(&plan)
-        );
-
-        self.execute_plan_and_build_stream(plan, tsd.query_id, addrs).await
-    }
-
-    /// Validate that addresses contain exactly one stage with the expected stage_id
-    fn validate_single_stage_addrs(&self, addrs: &Addrs, expected_stage_id: u64) -> Result<(), Status> {
-        if addrs.len() != 1 {
-            return Err(Status::internal(format!(
-                "Expected exactly one stage in addrs, got {}",
-                addrs.len()
-            )));
-        }
-        if !addrs.contains_key(&expected_stage_id) {
-            return Err(Status::internal(format!(
-                "No addresses found for stage_id {} in addrs",
-                expected_stage_id
-            )));
-        }
-        Ok(())
-    }
-
-    /// Execute a plan and build the response stream.
-    /// 
-    /// This function handles the common execution logic for both regular queries (which return data)
-    /// and EXPLAIN queries (which return plan information as text).
-    async fn execute_plan_and_build_stream(
-        &self,
-        plan: Arc<dyn ExecutionPlan>,
-        query_id: String,
-        addrs: Addrs,
-    ) -> Result<Response<crate::flight::DoGetStream>, Status> {
-        let mut ctx =
-            get_ctx().map_err(|e| Status::internal(format!("Could not create context {e:?}")))?;
-
-        add_ctx_extentions(&mut ctx, "proxy", &query_id, addrs, None)
-            .map_err(|e| Status::internal(format!("Could not add context extensions {e:?}")))?;
-
-        // TODO: revisit this to allow for consuming a partitular partition
-        trace!("calling execute plan");
-        let partition = 0;
-        let stream = plan
-            .execute(partition, ctx.task_ctx())
-            .map_err(|e| {
-                Status::internal(format!(
-                    "Error executing plan for query_id {} partition {}: {e:?}",
-                    query_id, partition
-                ))
-            })?
-            .map_err(|e| FlightError::ExternalError(Box::new(e)));
-
-        let out_stream = FlightDataEncoderBuilder::new()
-            .build(stream)
-            .map_err(move |e| Status::internal(format!("Unexpected error building stream {e:?}")));
-
-        Ok(Response::new(Box::pin(out_stream)))
     }
 }
 
@@ -473,9 +67,9 @@ impl FlightSqlHandler for DfRayProxyHandler {
         let is_explain = is_explain_query(&query.query);
 
         if is_explain {
-            self.handle_explain_request(&query.query).await
+            self.flight_handler.handle_explain_request(&query.query).await
         } else {
-            self.handle_query_request(&query.query).await
+            self.flight_handler.handle_query_request(&query.query).await
         }
     }
 
@@ -496,9 +90,9 @@ impl FlightSqlHandler for DfRayProxyHandler {
         debug!("request for ticket: {:?} from {}", tsd, remote_addr);
 
         if tsd.explain_data.is_some() {
-            self.handle_explain_statement_execution(tsd, &remote_addr).await
+            self.flight_handler.handle_explain_statement_execution(tsd, &remote_addr).await
         } else {
-            self.handle_regular_statement_execution(tsd, &remote_addr).await
+            self.flight_handler.handle_regular_statement_execution(tsd, &remote_addr).await
         }
     }
 }
@@ -614,408 +208,27 @@ impl DFRayProxyService {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use std::collections::HashMap;
-    
     // Test-specific imports
     use arrow_flight::{
         FlightDescriptor,
         Ticket,
-        decode::FlightRecordBatchStream,
         sql::{CommandStatementQuery, TicketStatementQuery},
     };
-    use futures::StreamExt;
     use prost::Message;
     use tonic::Request;
-
-    // //////////////////////////////////////////////////////////////
-    // Test helper functions
-    // //////////////////////////////////////////////////////////////
-
-    /// Create a test handler for testing - bypasses worker discovery initialization
-    fn create_test_handler() -> DfRayProxyHandler {
-        // Create the handler directly without calling new() to avoid worker discovery
-        // during test initialization.
-        DfRayProxyHandler {}
-    }
-
-    /// Create test worker addresses
-    fn create_test_addrs() -> Addrs {
-        let mut addrs = HashMap::new();
-        let mut stage_addrs = HashMap::new();
-        stage_addrs.insert(1u64, vec![
-            ("worker1".to_string(), "localhost:8001".to_string()),
-            ("worker2".to_string(), "localhost:8002".to_string()),
-        ]);
-        addrs.insert(1u64, stage_addrs);
-        addrs
-    }
-
-    /// Set up mock worker environment for testing
-    fn setup_mock_worker_env() {
-        let mock_addrs = vec![
-            ("mock_worker_1".to_string(), "localhost:9001".to_string()),
-            ("mock_worker_2".to_string(), "localhost:9002".to_string()),
-        ];
-        let mock_env_value = mock_addrs.iter()
-            .map(|(name, addr)| format!("{}/{}", name, addr))
-            .collect::<Vec<_>>()
-            .join(",");
-        std::env::set_var("DFRAY_WORKER_ADDRESSES", &mock_env_value);
-    }
-
-    /// Create TicketStatementData for EXPLAIN testing from QueryPlan
-    fn create_explain_ticket_statement_data(plans: QueryPlan) -> TicketStatementData {
-        let explain_data = plans.explain_data.map(|data| {
-            DistributedExplainExecNode {
-                schema: data.schema().as_ref().try_into().ok(),
-                logical_plan: data.logical_plan().to_string(),
-                physical_plan: data.physical_plan().to_string(),
-                distributed_plan: data.distributed_plan().to_string(),
-                distributed_stages: data.distributed_stages().to_string(),
-            }
-        });
-        
-        TicketStatementData {
-            query_id: plans.query_id,
-            stage_id: plans.final_stage_id,
-            stage_addrs: Some(plans.worker_addresses.into()),
-            schema: Some(plans.schema.as_ref().try_into().unwrap()),
-            explain_data,
-        }
-    }
-
-    /// Consume a DoGetStream and verify it contains expected EXPLAIN results
-    async fn verify_explain_stream_results(stream: crate::flight::DoGetStream) {
-        
-        // Convert the stream to a FlightRecordBatchStream and consume it
-        // Map Status errors to FlightError to match the expected stream type
-        let mapped_stream = stream.map(|result| {
-            result.map_err(|status| arrow_flight::error::FlightError::from(status))
-        });
-        let mut flight_stream = FlightRecordBatchStream::new_from_flight_data(mapped_stream);
-        
-        let mut batches = Vec::new();
-        while let Some(batch_result) = flight_stream.next().await {
-            let batch = batch_result.expect("Failed to get batch from stream");
-            batches.push(batch);
-        }
-        
-        // Verify we got exactly one batch with EXPLAIN results
-        assert_eq!(batches.len(), 1);
-        let batch = &batches[0];
-        
-        // Verify schema: should have 2 columns (plan_type, plan)
-        assert_eq!(batch.num_columns(), 2);
-        assert_eq!(batch.schema().field(0).name(), "plan_type");
-        assert_eq!(batch.schema().field(1).name(), "plan");
-        
-        // Verify we have 4 rows (logical_plan, physical_plan, distributed_plan, distributed_stages)
-        assert_eq!(batch.num_rows(), 4);
-        
-        // Verify the plan_type column contains the expected values
-        let plan_type_column = batch.column(0).as_any().downcast_ref::<arrow::array::StringArray>().unwrap();
-        assert_eq!(plan_type_column.value(0), "logical_plan");
-        assert_eq!(plan_type_column.value(1), "physical_plan");
-        assert_eq!(plan_type_column.value(2), "distributed_plan");
-        assert_eq!(plan_type_column.value(3), "distributed_stages");
-        
-        // Verify the plan column contains actual plan content
-        let plan_column = batch.column(1).as_any().downcast_ref::<arrow::array::StringArray>().unwrap();
-        assert!(plan_column.value(0).contains("Projection: Int64(1) AS test_col"));
-        assert!(plan_column.value(1).contains("ProjectionExec"));
-        assert!(plan_column.value(2).contains("RayStageExec"));
-        assert!(plan_column.value(3).contains("Stage 0:"));
-    }
-
-    // //////////////////////////////////////////////////////////////
-    // Unit tests for helper functions
-    // //////////////////////////////////////////////////////////////
-
-    #[test]
-    fn test_validate_single_stage_addrs_success() {
-        let handler = create_test_handler();
-        let addrs = create_test_addrs();
-
-        // Should succeed with valid single stage
-        let result = handler.validate_single_stage_addrs(&addrs, 1);
-        assert!(result.is_ok());
-    }
-
-    #[test]
-    fn test_validate_single_stage_addrs_multiple_stages() {
-        let handler = create_test_handler();
-        
-        // Create addresses with multiple stages
-        let mut addrs = HashMap::new();
-        let mut stage1_addrs = HashMap::new();
-        stage1_addrs.insert(1u64, vec![("worker1".to_string(), "localhost:8001".to_string())]);
-        let mut stage2_addrs = HashMap::new();
-        stage2_addrs.insert(2u64, vec![("worker2".to_string(), "localhost:8002".to_string())]);
-        addrs.insert(1u64, stage1_addrs);
-        addrs.insert(2u64, stage2_addrs);
-
-        // Should fail with multiple stages
-        let result = handler.validate_single_stage_addrs(&addrs, 1);
-        assert!(result.is_err());
-        
-        if let Err(status) = result {
-            assert!(status.message().contains("Expected exactly one stage"));
-        }
-    }
-
-    #[test]
-    fn test_validate_single_stage_addrs_wrong_stage_id() {
-        let handler = create_test_handler();
-        let addrs = create_test_addrs(); // Contains stage_id 1
-
-        // Should fail when looking for non-existent stage
-        let result = handler.validate_single_stage_addrs(&addrs, 999);
-        assert!(result.is_err());
-        
-        if let Err(status) = result {
-            assert!(status.message().contains("No addresses found for stage_id 999"));
-        }
-    }
-
-    #[test]
-    fn test_validate_single_stage_addrs_empty() {
-        let handler = create_test_handler();
-        let addrs: Addrs = HashMap::new();
-
-        // Should fail with empty addresses
-        let result = handler.validate_single_stage_addrs(&addrs, 1);
-        assert!(result.is_err());
-        
-        if let Err(status) = result {
-            assert!(status.message().contains("Expected exactly one stage"));
-        }
-    }
-
-
-    // //////////////////////////////////////////////////////////////
-    // Core function tests 
-    // //////////////////////////////////////////////////////////////
-
-    #[tokio::test]
-    async fn test_prepare_query_base() {
-        let handler = create_test_handler();
-        
-        // Test with a simple SELECT query without the need to read any table
-        let sql = "SELECT 1 as test_col";
-        let result = handler.prepare_query_base(sql, "TEST").await;
-        
-        if result.is_ok() {
-            let query_plan_base = result.unwrap();
-            // verify all fields have values
-            assert!(!query_plan_base.query_id.is_empty());
-            assert!(!query_plan_base.distributed_stages.is_empty());
-            assert!(!query_plan_base.physical_plan.schema().fields().is_empty());
-            // logical plan of select 1 on empty relation
-            assert_eq!(query_plan_base.logical_plan.to_string(), "Projection: Int64(1) AS test_col\n  EmptyRelation");
-            // physical plan of select 1 on empty releation is ProjectionExec
-            assert_eq!(query_plan_base.physical_plan.name(), "ProjectionExec");
-        } else {
-            // If worker discovery fails, we expect a specific error
-            let error_msg = format!("{:?}", result.unwrap_err());
-            assert!(error_msg.contains("worker") || error_msg.contains("address"));
-        }
-    }
-
-    #[tokio::test]
-    async fn test_prepare_explain() {
-        let handler = create_test_handler();
-        
-        // Test with a simple EXPLAIN query
-        let sql = "EXPLAIN SELECT 1 as test_col";
-        let query_plan = handler.prepare_explain(sql).await.unwrap();
-        
-        // EXPLAIN queries should work even without worker discovery since they use dummy addresses
-        assert!(!query_plan.query_id.is_empty());
-        assert_eq!(query_plan.final_stage_id, 0);
-        assert!(query_plan.explain_data.is_some());
-        
-        // Verify content of the explain data
-        let explain_data = query_plan.explain_data.unwrap();
-        assert_eq!(explain_data.logical_plan(), "Projection: Int64(1) AS test_col\n  EmptyRelation");
-        assert_eq!(explain_data.physical_plan(), "ProjectionExec: expr=[1 as test_col]\n  PlaceholderRowExec\n");
-        assert_eq!(explain_data.distributed_plan(), 
-            "RayStageExec[0] (output_partitioning=UnknownPartitioning(1))\n  ProjectionExec: expr=[1 as test_col]\n    PlaceholderRowExec\n");
-        assert_eq!(explain_data.distributed_stages(), 
-            "Stage 0:\n  Partition Groups: [[0]]\n  Full Partitions: false\n  Plan:\n    MaxRowsExec[max_rows=8192]\n      CoalesceBatchesExec: target_batch_size=8192\n        ProjectionExec: expr=[1 as test_col]\n          PlaceholderRowExec\n");
-        
-        // Should have explain schema (plan_type, plan columns)
-        assert_eq!(query_plan.schema.fields().len(), 2);
-        assert_eq!(query_plan.schema.field(0).name(), "plan_type");
-        assert_eq!(query_plan.schema.field(1).name(), "plan");
-        println!("✓ prepare_explain_query succeeded with proper structure");
-          
-    }
-
-    #[tokio::test]
-    async fn test_prepare_explain_invalid_input() {
-        let handler = create_test_handler();
-        
-        // Test with EXPLAIN ANALYZE (should fail)
-        let sql = "EXPLAIN ANALYZE SELECT 1";
-        let result = handler.prepare_explain(sql).await;
-        assert!(result.is_err());
-        let error_msg = format!("{:?}", result.unwrap_err());
-        assert!(error_msg.contains("prepare_explain called with non-EXPLAIN query"));
-        
-        // Test with non-EXPLAIN query (should fail)
-        let sql = "SELECT 1";
-        let result = handler.prepare_explain(sql).await;
-        assert!(result.is_err());
-        let error_msg = format!("{:?}", result.unwrap_err());
-        assert!(error_msg.contains("prepare_explain called with non-EXPLAIN query"));
-    }
-
-    // NOTE: This test is ignored because prepare_query() requires actual worker communication.
-    // 
-    // 🔍 Root Cause Analysis:
-    // The issue is NOT with mock worker setup - that works perfectly. The problem is in the 
-    // distribute_stages() retry logic:
-    //
-    // 1. ✅ Mock workers are set up correctly: ["mock_worker_1/localhost:9001", "mock_worker_2/localhost:9002"]
-    // 2. ✅ get_worker_addresses() successfully returns: [("mock_worker_1", "localhost:9001"), ("mock_worker_2", "localhost:9002")]
-    // 3. ✅ distribute_stages() receives workers and creates HashMap: {"mock_worker_1": "localhost:9001", "mock_worker_2": "localhost:9002"}
-    // 4. ❌ try_distribute_stages() attempts to create Flight client connections to mock workers (which don't exist)
-    // 5. ❌ Each connection fails, returning WorkerCommunicationError("mock_worker_X")
-    // 6. ❌ Retry logic removes "failed" workers: first removes mock_worker_2, then mock_worker_1
-    // 7. ❌ After 3 retries, workers HashMap is empty: {}
-    // 8. ❌ assign_to_workers() panics when trying to access worker_addrs[0] on empty list
-    //
-    // 💡 Solutions for proper testing:
-    // - Mock the Flight client layer (complex, requires significant refactoring)
-    // - Create a test-only version of distribute_stages() that skips communication
-    // - Refactor the architecture to use dependency injection for better testability
-    // - Use integration tests with actual worker processes instead of unit tests
-    //
-    // For now, we focus on testing the individual components that don't require worker communication.
-    #[tokio::test]
-    #[ignore]
-    async fn test_prepare_query() {
-        setup_mock_worker_env();
-        let handler = create_test_handler();
-        
-        // Test with a simple SELECT query
-        let sql = "SELECT 1 as test_col, 'hello' as text_col";
-        let result = handler.prepare_query(sql).await;
-        
-        match result {
-            Ok(query_plan) => {
-                assert!(query_plan.explain_data.is_none());
-                assert!(!query_plan.query_id.is_empty());
-                assert!(!query_plan.worker_addresses.is_empty());
-                assert_eq!(query_plan.schema.fields().len(), 2);
-                assert_eq!(query_plan.schema.field(0).name(), "test_col");
-                assert_eq!(query_plan.schema.field(1).name(), "text_col");
-                println!("✓ prepare_query succeeded with proper structure");
-            }
-            Err(e) => {
-                let error_msg = format!("{:?}", e);
-                assert!(
-                    error_msg.contains("worker") || 
-                    error_msg.contains("address") || 
-                    error_msg.contains("DFRAY_WORKER") ||
-                    error_msg.contains("index out of bounds"),
-                    "Unexpected error type: {}", error_msg
-                );
-                println!("✓ prepare_query failed with expected worker discovery error: {}", error_msg);
-            }
-        }
-    }
-
-    // //////////////////////////////////////////////////////////////
-    // EXPLAIN Flow Integration Tests
-    // //////////////////////////////////////////////////////////////
-
-    #[tokio::test]
-    async fn test_handle_explain_request() {
-        let handler = create_test_handler();
-        let query = "EXPLAIN SELECT 1 as test_col, 'hello' as text_col";
-        
-        let result = handler.handle_explain_request(query).await;
-        assert!(result.is_ok());
-        
-        let response = result.unwrap();
-        let flight_info = response.into_inner();
-        
-        // Verify FlightInfo structure
-        assert!(!flight_info.schema.is_empty());
-        assert_eq!(flight_info.endpoint.len(), 1);
-        assert!(flight_info.endpoint[0].ticket.is_some());
-        
-        // Verify that ticket has content (encoded TicketStatementData)
-        let ticket = flight_info.endpoint[0].ticket.as_ref().unwrap();
-        assert!(!ticket.ticket.is_empty());
-        
-        println!("✓ FlightInfo created successfully with {} schema bytes and ticket with {} bytes", 
-                 flight_info.schema.len(), ticket.ticket.len());
-    }
-
-    #[tokio::test]
-    async fn test_handle_explain_request_invalid_query() {
-        let handler = create_test_handler();
-        
-        // Test with EXPLAIN ANALYZE (should fail)
-        let query = "EXPLAIN ANALYZE SELECT 1";
-        let result = handler.handle_explain_request(query).await;
-        assert!(result.is_err());
-        
-        let error = result.unwrap_err();
-        assert_eq!(error.code(), tonic::Code::Internal);
-        assert!(error.message().contains("Could not prepare EXPLAIN query"));
-    }
-
-    #[tokio::test]
-    async fn test_handle_explain_statement_execution() {
-        let handler = create_test_handler();
-        
-        // First prepare an EXPLAIN query to get the ticket data structure
-        let query = "EXPLAIN SELECT 1 as test_col";
-        let plans = handler.prepare_explain(query).await.unwrap();
-        
-        // Create the TicketStatementData that would be sent to do_get_statement
-        let tsd = create_explain_ticket_statement_data(plans);
-        
-        // Test the execution
-        let result = handler.handle_explain_statement_execution(tsd, "test_remote").await;
-        assert!(result.is_ok());
-        
-        let response = result.unwrap();
-        let stream = response.into_inner();
-        
-        // Use shared verification function
-        verify_explain_stream_results(stream).await;
-    }
-
-    #[tokio::test]
-    async fn test_handle_explain_statement_execution_missing_explain_data() {
-        let handler = create_test_handler();
-        
-        // Create TicketStatementData without explain_data (should fail)
-        let tsd = TicketStatementData {
-            query_id: "test_query".to_string(),
-            stage_id: 0,
-            stage_addrs: None,
-            schema: None,
-            explain_data: None,
-        };
-        
-        let result = handler.handle_explain_statement_execution(tsd, "test_remote").await;
-        assert!(result.is_err());
-        
-        if let Err(error) = result {
-            assert_eq!(error.code(), tonic::Code::Internal);
-            assert!(error.message().contains("No explain_data in TicketStatementData"));
-        }
-    }
+    
+    use crate::{
+        test_utils::explain_test_helpers::{
+            create_explain_ticket_statement_data, 
+            create_test_proxy_handler, 
+            verify_explain_stream_results
+        },
+    };
 
     #[tokio::test]
     async fn test_get_flight_info_statement_explain() {
         
-        let handler = create_test_handler();
+        let handler = create_test_proxy_handler();
         
         // Test EXPLAIN query
         let command = CommandStatementQuery {
@@ -1046,11 +259,11 @@ mod tests {
     #[tokio::test]
     async fn test_do_get_statement_explain() {
         
-        let handler = create_test_handler();
+        let handler = create_test_proxy_handler();
         
         // First prepare an EXPLAIN query to get proper ticket data
         let query = "EXPLAIN SELECT 1 as test_col";
-        let plans = handler.prepare_explain(query).await.unwrap();
+        let plans = handler.flight_handler.planner.prepare_explain(query).await.unwrap();
         
         let tsd = create_explain_ticket_statement_data(plans);
         
@@ -1072,11 +285,11 @@ mod tests {
 
     #[tokio::test]
     async fn test_compare_explain_flight_info_responses() {
-        let handler = create_test_handler();
+        let handler = create_test_proxy_handler();
         let query = "EXPLAIN SELECT 1 as test_col";
         
         // Get FlightInfo from handle_explain_request
-        let result1 = handler.handle_explain_request(query).await.unwrap();
+        let result1 = handler.flight_handler.handle_explain_request(query).await.unwrap();
         let flight_info1 = result1.into_inner();
         
         // Get FlightInfo from get_flight_info_statement
@@ -1104,4 +317,8 @@ mod tests {
         println!("  - Endpoints: {} vs {}", flight_info1.endpoint.len(), flight_info2.endpoint.len());
         println!("  - Ticket bytes: {} vs {}", ticket1.ticket.len(), ticket2.ticket.len());
     }
+
+    // TODO: Add tests for regular (non-explain) queries
+    // We might need to create integration or end-to-end test infrastructure for this because
+    // they need workers
 }
